requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
supabase>=2.0.0
reportlab>=4.0.0
google-cloud-aiplatform>=1.35.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain_community